fix issue where the millis would roll over in only 36 hours, causing the RGB LED to stop blinking

+++ platform/MCU/NRF51/NRF51_StdPeriph_Driver/inc/libraries/timer/app_timer.h
@@ -129,7 +129,7 @@ typedef uint32_t (*app_timer_evt_schedule_func_t) (app_timer_timeout_handler_t t
 
 
 /**@brief Integer for keepping track of the RTC overflow events. */
-volatile uint8_t RTC_OVERFLOW_COUNT;
+volatile uint32_t RTC_OVERFLOW_COUNT;
 
 
 /**@brief Timer modes. */
