+++ platform/MCU/NRF51/SPARK_Firmware_Driver/src/hw_config.c
@@ -39,13 +39,12 @@ uint32_t External_Flash_Start_Address = 0;
 
 uint32_t system_millis(void)
 {
-//    return system_milliseconds;
-    return (( ((RTC_OVERFLOW_COUNT << 24) | NRF_RTC1->COUNTER) * 1.0 ) / (APP_TIMER_CLOCK_FREQ * 1.0)) * 1000.0;
+    return (( (((uint64_t)RTC_OVERFLOW_COUNT << 24) | (uint64_t)NRF_RTC1->COUNTER) * 1.0 ) / (APP_TIMER_CLOCK_FREQ * 1.0)) * 1000.0;
 }
 
 uint32_t system_micros(void)
 {
-    return (( ((RTC_OVERFLOW_COUNT << 24) | NRF_RTC1->COUNTER) * 1.0 ) / (APP_TIMER_CLOCK_FREQ * 1.0)) * 1000000.0;
+    return (( (((uint64_t)RTC_OVERFLOW_COUNT << 24) | (uint64_t)NRF_RTC1->COUNTER) * 1.0 ) / (APP_TIMER_CLOCK_FREQ * 1.0)) * 1000000.0;
 }
 
 /**@brief Function for error handling, which is called when an error has occurred.
